+++ packages/storybook/examples/core/ColumnView/index.js
@@ -1,0 +1,12 @@
+import { storiesOf } from '@storybook/react';
+import { withInfo } from '@storybook/addon-info';
+
+import ColumnView from '@ichef/gypcrete/src/ColumnView';
+import getPropTables from 'utils/getPropTables';
+import BasicUsage from './BasicUsage';
+
+storiesOf('@ichef/gypcrete|ColumnView', module)
+    .add('basic usage',
+        withInfo()(BasicUsage)
+    )
+    .add('props', getPropTables([ColumnView]));